backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
openai>=1.50.0
requests==2.31.0
alembic==1.14.0
email-validator==2.1.0
slowapi==0.1.9
bcrypt>=4.0.0
resend>=2.0.0
plaid-python>=14.0.0
sentry-sdk[fastapi]>=2.0.0
